requirements.txt
gradio>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=9.5.0
matplotlib>=3.7.0
huggingface-hub>=0.16.0
spaces>=0.25.0
transformers>=4.30.0
timm>=0.9.0
gradio_imageslider
open3d
scikit-learn
git+https://github.com/EasternJournalist/utils3d.git@c5daf6f6c244d251f252102d09e9b7bcef791a38
click           # ==8.1.7
trimesh         # ==4.5.1